+++ CHANGELOG.md
@@ -13,6 +13,7 @@ omnibus-gitlab repository.
 - Added nginx options to enable 2-way SSL client authentication (Oliver Hernandez) c51085
 - Upgrade libicu to 57.1 f58a4b15
 - Upgrade Nginx to 1.10.1 67a0bd0
+- Allow configuration of the authorized_keys file location used by gitlab-shell
 - Upgrade omnibus to 5.4.0 7bac2
 - Add configuration that allows disabling of db migrations (Jason Plum) a50d09
 - Initial support for Redis Sentinel 267ace